arm64: Track per-cpu MPIDR in arch_cpu_info
Apple's interrupt controller needs to know the MPIDR of the
CPU you want to send an ICI to

Reviewed-on: https://review.haiku-os.org/c/haiku/+/11124
Tested-by: Commit checker robot <no-reply+buildbot@haiku-os.org>

+++ headers/private/kernel/arch/arm64/arch_cpu.h
@@ -126,7 +126,7 @@ namespace BKernel {
 
 
 typedef struct arch_cpu_info {
-	uint32						mpidr;
+	uint64						mpidr;
 	BKernel::Thread*			last_vfp_user;
 } arch_cpu_info;
 #endif
